+++ src/main/java/com/gregtechceu/gtceu/api/sync_system/ClassSyncData.java
@@ -6,6 +6,8 @@ import com.gregtechceu.gtceu.api.sync_system.annotations.SaveField;
 import com.gregtechceu.gtceu.api.sync_system.annotations.SyncToClient;
 import com.gregtechceu.gtceu.api.sync_system.data_transformers.ValueTransformer;
 import com.gregtechceu.gtceu.api.sync_system.data_transformers.ValueTransformers;
+import com.gregtechceu.gtceu.api.sync_system.managed.ISyncAnnotated;
+import com.gregtechceu.gtceu.api.sync_system.managed.ISyncManaged;
 
 import it.unimi.dsi.fastutil.objects.ObjectArrayList;
 import it.unimi.dsi.fastutil.objects.ObjectOpenHashSet;
