+++ src/main/java/com/gregtechceu/gtceu/api/sync_system/SyncDataHolder.java
@@ -2,6 +2,7 @@ package com.gregtechceu.gtceu.api.sync_system;
 
 import com.gregtechceu.gtceu.GTCEu;
 import com.gregtechceu.gtceu.api.sync_system.data_transformers.ValueTransformer;
+import com.gregtechceu.gtceu.api.sync_system.managed.ISyncManaged;
 
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.Tag;
