+++ src/main/java/com/gregtechceu/gtceu/api/sync_system/managed/package-info.java
@@ -1,0 +1,4 @@
+@NotNullByDefault
+package com.gregtechceu.gtceu.api.sync_system.managed;
+
+import org.jetbrains.annotations.NotNullByDefault;
